commit: merge encode into id module
